Move `WorkspacesPreview` and friends to its own folder
