+++ client/web/src/enterprise/batches/create/workspaces-preview/WorkspacesPreview.mock.ts
@@ -1,4 +1,4 @@
-import { BatchSpecWorkspaceResolutionState, WorkspaceResolutionStatusResult } from '../../../graphql-operations'
+import { BatchSpecWorkspaceResolutionState, WorkspaceResolutionStatusResult } from '../../../../graphql-operations'
 
 export const mockWorkspaceResolutionStatus = (
     status: BatchSpecWorkspaceResolutionState,
