+++ client/web/src/enterprise/batches/create/workspaces-preview/WorkspacesPreview.module.scss
